// client/model/getManyWanFormRequestResponseDto.ts
/**
 * Draas
 * Draas OAS Documentation
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WanFormRequest } from './wanFormRequest';


export interface GetManyWanFormRequestResponseDto { 
    totalPages: number;
    count: number;
    total: number;
    page: number;
    pageCount: number;
    data: Array<WanFormRequest>;
}